requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==4.2.26
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.3
whitenoise==6.11.0
